% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.Ts}
\alias{get.Ts}
\title{gets surface water temperatures}
\usage{
get.Ts(data, s.range = c(0, 1))
}
\arguments{
\item{data}{Object of class data.frame}

\item{s.range}{a numeric vector of length=2 with the range for depth measurements to still be considered 'surface'}
}
\value{
An object of class data.frame
}
\description{
grabs best available data for surface water temperature
}
\author{
Jordan S. Read
}
\seealso{
\link{has.vars}
\link{get.vars}
\link{rmv.vars}
}
\keyword{methods}

