% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.meta.R
\name{load.meta}
\alias{load.meta}
\title{Loads a metadata file from the specified path}
\usage{
load.meta(fPath)
}
\arguments{
\item{fPath}{The file path as a string}
}
\value{
A list with the metadata parsed from the file.
}
\description{
Parses a formatted metadata file. Useful for site-specific metadata 
that is not contained in the timeseries files.
}
\seealso{
\link{load.ts}
\link{load.all.data}
}
\author{
Luke A Winslow
}
\keyword{IO}
\keyword{file}
