% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metab.kalman}
\alias{metab.kalman}
\title{Estimate metabolism using a Kalman filter}
\usage{
metab.kalman(do.obs, do.sat, k.gas, z.mix, irr, wtr, ...)
}
\arguments{
\item{do.obs}{Vector of dissovled oxygen concentration observations, mg L^-1}

\item{do.sat}{Vector of dissolved oxygen saturation values based on water temperature. Calculate using \link{o2.at.sat}}

\item{k.gas}{Vector of kGAS values calculated from any of the gas flux models
(e.g., \link{k.cole}) and converted to kGAS using \link{k600.2.kGAS}}

\item{z.mix}{Vector of mixed-layer depths in meters. To calculate, see \link{ts.meta.depths}}

\item{irr}{Vector of photosynthetically active radiation in umoles/m2/s}

\item{wtr}{Vector of water temperatures in deg C. Used in scaling respiration with temperature}

\item{priors}{Parameter priors supplied as a named list (example: c("gppMu"=0, "gppSig2"=1E5, "rMu"=0, "rSig2"=1E5, "kSig2"=NA))}

\item{...}{additional arguments to be passed to \link{optim}}
}
\value{
A named list of parameter estimates.
\item{smoothDO}{smoothed time series of oxygen concentration, from Kalman smoother}
\item{params}{parameters estimated by the Kalman filter}
\item{metab}{daily metabolism estimates in mg O2 / L / day}
}
\description{
Uses a Kalman filter to fit parameters relating irr to GPP, log(wtr) to R, process error, and observation error.
Also provides a smoothed time series of oxygen concentration.
}
\examples{
doobs <- load.ts(system.file('extdata',
                            'Sparkling.doobs', package="LakeMetabolizer"))
wtr <- load.ts(system.file('extdata',
                          'Sparkling.wtr', package="LakeMetabolizer"))
wnd <- load.ts(system.file('extdata',
                          'Sparkling.wnd', package="LakeMetabolizer"))
irr <- load.ts(system.file('extdata',
                          'Sparkling.par', package="LakeMetabolizer"))

#Subset a day
Sys.setenv(TZ='GMT')
mod.date <- as.POSIXct('2009-08-12')
doobs <- doobs[trunc(doobs$datetime, 'day') == mod.date, ]
wtr <- wtr[trunc(wtr$datetime, 'day') == mod.date, ]
wnd <- wnd[trunc(wnd$datetime, 'day') == mod.date, ]
irr <- irr[trunc(irr$datetime, 'day') == mod.date, ]

k600 <- k.cole.base(wnd[,2])
k.gas <- k600.2.kGAS.base(k600, wtr[,3], 'O2')
do.sat <- o2.at.sat.base(wtr[,3], altitude=300)

metab.kalman(irr=irr[,2], z.mix=rep(1, length(k.gas)),
            do.sat=do.sat, wtr=wtr[,2],
            k.gas=k.gas, do.obs=doobs[,2])
}
\author{
Ryan Batt, Luke A. Winslow
}
\references{
Batt, Ryan D. and Stephen R. Carpenter. 2012. \emph{Free-water lake metabolism:
addressing noisy time series with a Kalman filter}. Limnology and Oceanography: Methods 10: 20-30. doi: 10.4319/lom.2012.10.20
}
\seealso{
\link{temp.kalman}, \link{watts.in}, \link{metab}, \link{metab.mle}, \link{metab.bookkeep}, \link{metab.kalman}
}

