% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metab.mle}
\alias{metab.mle}
\title{Metabolism model based on a maximum likelihood parameter estimation framework.}
\usage{
metab.mle(do.obs, do.sat, k.gas, z.mix, irr, wtr, ...)
}
\arguments{
\item{do.obs}{Vector of dissolved oxygen concentration observations, mg L^-1}

\item{do.sat}{Vector of dissolved oxygen saturation values based on water temperature. Calculate using \link{o2.at.sat}}

\item{k.gas}{Vector of kGAS values calculated from any of the gas flux models
(e.g., \link{k.cole}) and converted to kGAS using \link{k600.2.kGAS}}

\item{z.mix}{Vector of mixed-layer depths in meters. To calculate, see \link{ts.meta.depths}}

\item{irr}{Vector of photosynthetically active radiation in umoles/m2/s}

\item{wtr}{Vector of water temperatures in deg C. Used in scaling respiration with temperature}

\item{...}{additional arguments to be passed}
}
\value{
A named list of parameter estimates.
\item{GPP}{Estimated Gross Primary Productivity}
\item{R}{Estimated ecosystem respiration}
}
\description{
This function runs the maximum likelihood metabolism model on the supplied gas concentration and other supporting data. This is a
common approach that allows for the concurrent estimation of metabolism paramters from a timeseries.
}
\examples{
library(rLakeAnalyzer)
doobs = load.ts(system.file('extdata',
                           'Sparkling.doobs', package="LakeMetabolizer"))
wtr = load.ts(system.file('extdata',
                         'Sparkling.wtr', package="LakeMetabolizer"))
wnd = load.ts(system.file('extdata',
                         'Sparkling.wnd', package="LakeMetabolizer"))
irr = load.ts(system.file('extdata',
                         'Sparkling.par', package="LakeMetabolizer"))

#Subset a day
mod.date = as.POSIXct('2009-08-12')
doobs = doobs[trunc(doobs$datetime, 'day') == mod.date, ]
wtr = wtr[trunc(wtr$datetime, 'day') == mod.date, ]
wnd = wnd[trunc(wnd$datetime, 'day') == mod.date, ]
irr = irr[trunc(irr$datetime, 'day') == mod.date, ]
z.mix = ts.thermo.depth(wtr)

k600 = k.cole.base(wnd[,2])
k.gas = k600.2.kGAS.base(k600, wtr[,3], 'O2')
do.sat = o2.at.sat.base(wtr[,3], altitude=300)

metab.mle(doobs[,2], do.sat, k.gas, z.mix[,2], irr[,2], wtr[,3])
}
\author{
Luke A Winslow, Ryan Batt, GLEON Fellows
}
\references{
Solomon CT, DA Bruesewitz, DC Richardson, KC Rose, MC Van de Bogert, PC Hanson, TK Kratz, B Larget,
R Adrian, B Leroux Babin, CY Chiu, DP Hamilton, EE Gaiser, S Hendricks, V Istvanovics, A Laas, DM O'Donnell,
ML Pace, E Ryder, PA Staehr, T Torgersen, MJ Vanni, KC Weathers, G Zhuw 2013.
\emph{Ecosystem Respiration: Drivers of Daily Variability and Background Respiration in Lakes around the Globe}.
Limnology and Oceanograph 58 (3): 849:866. doi:10.4319/lo.2013.58.3.0849.
}
\seealso{
\link{metab.bayesian}, \link{metab.bookkeep}, \link{metab.ols}
}

