% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metab.ols}
\alias{metab.ols}
\title{Metabolism model based on a ordinary least squares parameter estimation framework.}
\usage{
metab.ols(do.obs, do.sat, k.gas, z.mix, irr, wtr, ...)
}
\arguments{
\item{do.obs}{Vector of dissolved oxygen concentration observations, mg L^-1}

\item{do.sat}{Vector of dissolved oxygen saturation values based on water temperature. Calculate using \link{o2.at.sat}}

\item{k.gas}{Vector of kGAS values calculated from any of the gas flux models
(e.g., \link{k.cole}) and converted to kGAS using \link{k600.2.kGAS}}

\item{z.mix}{Vector of mixed-layer depths in meters. To calculate, see \link{ts.meta.depths}}

\item{irr}{Vector of photosynthetically active radiation in umoles/m2/s}

\item{wtr}{Vector of water temperatures in deg C. Used in scaling respiration with temperature}

\item{priors}{Parameter priors supplied as a named list (example: c("gppMu"=0, "gppSig2"=1E5, "rMu"=0, "rSig2"=1E5, "kSig2"=NA))}

\item{...}{additional arguments to be passed}
}
\value{
A named list of parameter estimates.
\item{GPP}{Estimated Gross Primary Productivity}
\item{R}{Estimated ecosystem Respiration}
\item{NEP}{Net Ecosystem Productivity}
}
\description{
This function runs the ordinary least squares metabolism model on the supplied gas concentration and other supporting data. This is a
common approach that allows for the concurrent estimation of metabolism paramters from a timeseries.
}
\examples{
library(rLakeAnalyzer)
doobs = load.ts(system.file('extdata',
                           'Sparkling.doobs', package="LakeMetabolizer"))
wtr = load.ts(system.file('extdata',
                         'Sparkling.wtr', package="LakeMetabolizer"))
wnd = load.ts(system.file('extdata',
                         'Sparkling.wnd', package="LakeMetabolizer"))
irr = load.ts(system.file('extdata',
                         'Sparkling.par', package="LakeMetabolizer"))

#Subset a day
mod.date = as.POSIXct('2009-08-12')
doobs = doobs[trunc(doobs$datetime, 'day') == mod.date, ]
wtr = wtr[trunc(wtr$datetime, 'day') == mod.date, ]
wnd = wnd[trunc(wnd$datetime, 'day') == mod.date, ]
irr = irr[trunc(irr$datetime, 'day') == mod.date, ]
z.mix = ts.thermo.depth(wtr)

k600 = k.cole.base(wnd[,2])
k.gas = k600.2.kGAS.base(k600, wtr[,3], 'O2')
do.sat = o2.at.sat.base(wtr[,3], altitude=300)

metab.ole(doobs[,2], do.sat, k.gas, z.mix[,2], irr[,2], wtr[,3])
}
\author{
Luke A Winslow, Ryan Batt, GLEON Fellows
}
\seealso{
\link{metab.bayesian}, \link{metab.bookeep}, \link{metab.mle}, \link{metab}
}

