% Generated by roxygen2 (4.0.2): do not edit by hand
\name{par.to.sw}
\alias{par.to.sw}
\alias{par.to.sw.base}
\title{Convert PAR to shortwave}
\usage{
par.to.sw.base(par)

par.to.sw(data, par.col='par')
}
\arguments{
\item{data}{Object of class data.frame with column name 'par' (units umol/m^2/sec)}

\item{par.col}{String of alternative name for PAR column}

\item{par}{Numeric vector of PAR values}
}
\value{
#For par.to.sw

Object of class data.frame with column name 'sw' and other values from \code{ts.data}

#For par.to.sw.base

Numeric vector of shortwave values with units W/m^2
}
\description{
Returns incoming shortwave radiation by converting PAR measuremt.
}
\examples{
par <- 800
par.to.sw.base(par)
}
\author{
LakeMetabolizer
}
\references{
Britton, C. M., and J. D. Dodd. \emph{Relationships of photosynthetically active radiation and shortwave irradiance.}
Agricultural Meteorology 17, no. 1 (1976): 1-7.
}
\seealso{
\link{sw.to.par}
}
\keyword{math}
\keyword{methods}

