% Generated by roxygen2 (4.0.1): do not edit by hand
\name{var.indx}
\alias{var.indx}
\title{finds matching column names in data.frame}
\usage{
var.indx(data, var.name)
}
\arguments{
\item{data}{Object of class data.frame}

\item{var.name}{A character vector of names to find matches with \code{data}}
}
\value{
a boolean vector with same length as \code{var.names}
}
\description{
returns index of column matches for \code{data} according to header names matches with \code{var.names}.
}
\author{
Luke A. Winslow
}
\seealso{
\link{has.vars}
\link{get.vars}
\link{rmv.vars}
}
\keyword{methods}

