\name{get.vars}
\alias{get.vars}
\title{subsets data.frame according to header names}
\usage{
  get.vars(data, var.names)
}
\arguments{
  \item{\code{data}}{Object of class data.frame}

  \item{\code{var.names}}{A character vector of names to
  get from \code{data}}
}
\value{
  An object of class data.frame
}
\description{
  subsets \code{data} according to header names
}
\author{
  Luke A. Winslow
}
\seealso{
  \link{has.vars} \link{rmv.vars}
}
\keyword{methods}

