% Generated by roxygen2 (4.0.1): do not edit by hand
\name{has.vars}
\alias{has.vars}
\title{tests data.frame for column names}
\usage{
has.vars(data, var.names)
}
\arguments{
\item{\code{data}}{Object of class data.frame}

\item{\code{var.names}}{A character vector of names to test against \code{data}}
}
\value{
a boolean vector of same length as \code{var.names}
}
\description{
tests \code{data} for data column names
}
\author{
Luke A. Winslow
}
\seealso{
\link{get.vars}
\link{rmv.vars}
}
\keyword{methods}

