% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.night}
\alias{is.night}
\title{determines if measurement was taken during the night}
\usage{
is.day(datetimes, lat)
}
\arguments{
\item{\code{datetimes}}{Vector of dates as \code{POSIXct} or \code{POSIXlt} (see \code{\link{DateTimeClasses}}) format}

\item{\code{lat}}{Single latitude value of site. South should be negative, north positive}
}
\value{
a boolean vector of same length as \code{datetimes}
}
\description{
determines if measurement was taken during the nighttime
}
\author{
Luke A. Winslow
}
\seealso{
\link{is.day}
\link{sun.rise.set}
}
\keyword{methods}

