% Generated by roxygen2 (4.0.1): do not edit by hand
\name{par.to.sw}
\alias{par.to.sw}
\title{Convert PAR to shortwave}
\usage{
calc.lw.net(ts.data, lat, atm.press)

calc.lw.net.base(dateTime, sw, Ts, lat, atm.press, airT, RH)
}
\arguments{
\item{\code{ts.data}}{Object of class data.frame with column name 'par'}
}
\value{
Object of class data.frame with column name 'sw' and other values from \code{ts.data}
}
\description{
Returns incoming shortwave radiation by converting PAR measuremt.
}
\examples{
par <- 800
par.to.sw.base(par)
}
\author{
LakeMetabolizer
}
\references{
Britton, C. M., and J. D. Dodd. \emph{Relationships of photosynthetically active radiation and shortwave irradiance.}
Agricultural Meteorology 17, no. 1 (1976): 1-7.
}
\seealso{
\link{sw.to.par}
}
\keyword{math}
\keyword{methods}

